website_summarizer/requirements.txt
langchain
faiss-cpu
openai
python-dotenv
selenium
beautifulsoup4
streamlit
tiktoken
PyPDF2
webdriver-manager
